+++ src/BabylonImGui/src/babylon_imgui/babylon_studio.cpp
@@ -8,7 +8,7 @@
 #include <babylon/interfaces/irenderable_scene_with_hud.h>
 #include <imgui.h>
 #include <imgui_internal.h>
-#include <imgui_runner_babylon/runner_babylon.h>
+#include <imgui_utils/imgui_runner_babylon/runner_babylon.h>
 #include <imgui_utils/icons_font_awesome_5.h>
 
 #include <babylon/babylon_imgui/babylon_studio_layout.h>
