+++ src/imgui_runner_demos/imgui_runner_demo_emscripten.cpp
@@ -1,5 +1,5 @@
 #include "example_gui.h"
-#include "imgui_runner/runner_emscripten.h"
+#include "imgui_utils/imgui_runner/runner_emscripten.h"
 
 int main()
 {
